Adds received jar files to the CLASSPATH
This feature allows us to use the jar files in the "submitted files" or
"execution files". If you add jar files in the execution files to be
used in the running process remember to add the jar files to the "Files
to keep when running" at the activity "Advanced settings".