WebService now accepts binary files using Base64
It seems that the WebService API does not accepts binary data.
For the requests "info", "open" and "save" a new field in files
data has been added.
The field is called "encoding" if encoding is equal to 1 indicate that
the file data is encoded in base64.
Notice that this encoded is used in files with known binary
file name extensions.