Reduces maximum memory used by ZipArchive
VPL uses the addFromString() method to add files to the ZIP file.
addFromString() keeps data in memory until the ZIP file is written. The new approach closes and creates a new ZipArchive object to reopen the ZIP file every time the total size of the added files reaches 64MB. This action removes the data in memory.
Notice that the previous version required as much memory in the server as the total data of the files in the ZIP.