Add script to launch Python Tutor in VPL environment
This script enables the execution of student code (Python, C, C++, Java and JavaScript) within Python Tutor, leveraging the VPL web application feature. It uses a jail server as a relay to package and send the student's code to be executed and visualized  through Python Tutor. While this approach is not ideal, it is a straightforward solution.
Note: This may have provacy concerns.
Note: Python Tutor is limited to processing a single file.